requirements.txt
Flask==3.0.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==7.4.2
SQLAlchemy==2.0.30
Werkzeug==3.0.3
WTForms==3.1.2
